+++ geozero/src/gdal/gdal_reader.rs
@@ -156,6 +156,7 @@ impl GeozeroGeometry for Geometry {
 }
 
 #[cfg(test)]
+#[cfg(feature = "with-wkt")]
 mod test {
     use super::*;
     use crate::wkt::WktWriter;