+++ geozero/src/svg/mod.rs
@@ -1,5 +1,6 @@
 //! SVG conversions.
 mod svg;
+
 pub use svg::*;
 
 pub(crate) mod conversion {